refactor(cli): cleanups and clarification based on code review
