+++ cli/module_graph_2.rs
@@ -107,22 +107,6 @@ impl fmt::Display for GraphError {
 
 impl Error for GraphError {}
 
-/// A trait, implemented by `Graph` that provides the interfaces that the
-/// compiler ops require to be able to retrieve information about the graph.
-pub trait ModuleProvider {
-  /// Get the source for a given module specifier.  If the module is not part
-  /// of the graph, the result will be `None`.
-  fn get_source(&self, specifier: &ModuleSpecifier) -> Option<String>;
-  /// Given a string specifier and a referring module specifier, provide the
-  /// resulting module specifier and media type for the module that is part of
-  /// the graph.
-  fn resolve(
-    &self,
-    specifier: &str,
-    referrer: &ModuleSpecifier,
-  ) -> Result<(ModuleSpecifier, MediaType), AnyError>;
-}
-
 /// An enum which represents the parsed out values of references in source code.
 #[derive(Debug, Clone, Eq, PartialEq)]
 enum TypeScriptReference {
@@ -420,7 +404,6 @@ pub struct TranspileOptions {
 /// A dependency graph of modules, were the modules that have been inserted via
 /// the builder will be loaded into the graph.  Also provides an interface to
 /// be able to manipulate and handle the graph.
-
 #[derive(Debug)]
 pub struct Graph {
   build_info: BuildInfoMap,
@@ -476,6 +459,21 @@ impl Graph {
     Ok(())
   }
 
+  /// Get the source for a given module specifier.  If the module is not part
+  /// of the graph, the result will be `None`.
+  #[allow(unused)]
+  pub fn get_source(&self, specifier: &ModuleSpecifier) -> Option<String> {
+    if let Some(module) = self.modules.get(specifier) {
+      if let Ok(source) = module.source.to_string() {
+        Some(source)
+      } else {
+        None
+      }
+    } else {
+      None
+    }
+  }
+
   /// Verify the subresource integrity of the graph based upon the optional
   /// lockfile, updating the lockfile with any missing resources.  This will
   /// error if any of the resources do not match their lock status.
@@ -498,6 +496,63 @@ impl Graph {
     }
 
     Ok(())
+  }
+
+  /// Given a string specifier and a referring module specifier, provide the
+  /// resulting module specifier and media type for the module that is part of
+  /// the graph.
+  #[allow(unused)]
+  pub fn resolve(
+    &self,
+    specifier: &str,
+    referrer: &ModuleSpecifier,
+  ) -> Result<(ModuleSpecifier, MediaType), AnyError> {
+    if !self.modules.contains_key(referrer) {
+      return Err(MissingSpecifier(referrer.to_owned()).into());
+    }
+    let module = self.modules.get(referrer).unwrap();
+    if !module.dependencies.contains_key(specifier) {
+      return Err(
+        MissingDependency(referrer.to_owned(), specifier.to_owned()).into(),
+      );
+    }
+    let dependency = module.dependencies.get(specifier).unwrap();
+    // If there is a @deno-types pragma that impacts the dependency, then the
+    // maybe_type property will be set with that specifier, otherwise we use the
+    // specifier that point to the runtime code.
+    let resolved_specifier =
+      if let Some(type_specifier) = dependency.maybe_type.clone() {
+        type_specifier
+      } else if let Some(code_specifier) = dependency.maybe_code.clone() {
+        code_specifier
+      } else {
+        return Err(
+          MissingDependency(referrer.to_owned(), specifier.to_owned()).into(),
+        );
+      };
+    if !self.modules.contains_key(&resolved_specifier) {
+      return Err(
+        MissingDependency(referrer.to_owned(), resolved_specifier.to_string())
+          .into(),
+      );
+    }
+    let dep_module = self.modules.get(&resolved_specifier).unwrap();
+    // In the case that there is a X-TypeScript-Types or a triple-slash types,
+    // then the `maybe_types` specifier will be populated and we should use that
+    // instead.
+    let result = if let Some((_, types)) = dep_module.maybe_types.clone() {
+      if let Some(types_module) = self.modules.get(&types) {
+        (types, types_module.media_type)
+      } else {
+        return Err(
+          MissingDependency(referrer.to_owned(), types.to_string()).into(),
+        );
+      }
+    } else {
+      (resolved_specifier, dep_module.media_type)
+    };
+
+    Ok(result)
   }
 
   /// Transpile (only transform) the graph, updating any emitted modules
@@ -581,73 +636,6 @@ impl Graph {
     ]);
 
     Ok((stats, maybe_ignored_options))
-  }
-}
-
-impl<'a> ModuleProvider for Graph {
-  fn get_source(&self, specifier: &ModuleSpecifier) -> Option<String> {
-    if let Some(module) = self.modules.get(specifier) {
-      if let Ok(source) = module.source.to_string() {
-        Some(source)
-      } else {
-        None
-      }
-    } else {
-      None
-    }
-  }
-
-  fn resolve(
-    &self,
-    specifier: &str,
-    referrer: &ModuleSpecifier,
-  ) -> Result<(ModuleSpecifier, MediaType), AnyError> {
-    if !self.modules.contains_key(referrer) {
-      return Err(MissingSpecifier(referrer.to_owned()).into());
-    }
-    let module = self.modules.get(referrer).unwrap();
-    if !module.dependencies.contains_key(specifier) {
-      return Err(
-        MissingDependency(referrer.to_owned(), specifier.to_owned()).into(),
-      );
-    }
-    let dependency = module.dependencies.get(specifier).unwrap();
-    // If there is a @deno-types pragma that impacts the dependency, then the
-    // maybe_type property will be set with that specifier, otherwise we use the
-    // specifier that point to the runtime code.
-    let resolved_specifier =
-      if let Some(type_specifier) = dependency.maybe_type.clone() {
-        type_specifier
-      } else if let Some(code_specifier) = dependency.maybe_code.clone() {
-        code_specifier
-      } else {
-        return Err(
-          MissingDependency(referrer.to_owned(), specifier.to_owned()).into(),
-        );
-      };
-    if !self.modules.contains_key(&resolved_specifier) {
-      return Err(
-        MissingDependency(referrer.to_owned(), resolved_specifier.to_string())
-          .into(),
-      );
-    }
-    let dep_module = self.modules.get(&resolved_specifier).unwrap();
-    // In the case that there is a X-TypeScript-Types or a triple-slash types,
-    // then the `maybe_types` specifier will be populated and we should use that
-    // instead.
-    let result = if let Some((_, types)) = dep_module.maybe_types.clone() {
-      if let Some(types_module) = self.modules.get(&types) {
-        (types, types_module.media_type)
-      } else {
-        return Err(
-          MissingDependency(referrer.to_owned(), types.to_string()).into(),
-        );
-      }
-    } else {
-      (resolved_specifier, dep_module.media_type)
-    };
-
-    Ok(result)
   }
 }
 
@@ -756,6 +744,11 @@ impl GraphBuilder {
   /// lockfile can be provided, where if the sources in the graph do not match
   /// the expected lockfile, the method with error instead of returning the
   /// graph.
+  ///
+  /// TODO(@kitsonk) this should really be owned by the graph, but currently
+  /// the lockfile is behind a mutex in global_state, which makes it really
+  /// hard to not pass around as a reference, which if the Graph owned it, it
+  /// would need lifetime parameters and lifetime parameters are 😭
   pub fn get_graph(
     self,
     maybe_lockfile: &Option<Mutex<Lockfile>>,
@@ -769,11 +762,135 @@ impl GraphBuilder {
 mod tests {
   use super::*;
 
-  use crate::specifier_handler::tests::MockSpecifierHandler;
-
+  use deno_core::futures::future;
   use std::env;
+  use std::fs;
   use std::path::PathBuf;
   use std::sync::Mutex;
+
+  /// This is a testing mock for `SpecifierHandler` that uses a special file
+  /// system renaming to mock local and remote modules as well as provides
+  /// "spies" for the critical methods for testing purposes.
+  #[derive(Debug, Default)]
+  pub struct MockSpecifierHandler {
+    pub fixtures: PathBuf,
+    pub build_info: HashMap<ModuleSpecifier, TextDocument>,
+    pub build_info_calls: Vec<(ModuleSpecifier, EmitType, TextDocument)>,
+    pub cache_calls: Vec<(
+      ModuleSpecifier,
+      EmitType,
+      TextDocument,
+      Option<TextDocument>,
+    )>,
+    pub deps_calls: Vec<(ModuleSpecifier, DependencyMap)>,
+    pub types_calls: Vec<(ModuleSpecifier, String)>,
+    pub version_calls: Vec<(ModuleSpecifier, String)>,
+  }
+
+  impl MockSpecifierHandler {
+    fn get_cache(
+      &self,
+      specifier: ModuleSpecifier,
+    ) -> Result<CachedModule, AnyError> {
+      let specifier_text = specifier
+        .to_string()
+        .replace(":///", "_")
+        .replace("://", "_")
+        .replace("/", "-");
+      let specifier_path = self.fixtures.join(specifier_text);
+      let media_type =
+        match specifier_path.extension().unwrap().to_str().unwrap() {
+          "ts" => {
+            if specifier_path.to_string_lossy().ends_with(".d.ts") {
+              MediaType::Dts
+            } else {
+              MediaType::TypeScript
+            }
+          }
+          "tsx" => MediaType::TSX,
+          "js" => MediaType::JavaScript,
+          "jsx" => MediaType::JSX,
+          _ => MediaType::Unknown,
+        };
+      let source =
+        TextDocument::new(fs::read(specifier_path)?, Option::<&str>::None);
+
+      Ok(CachedModule {
+        source,
+        specifier,
+        media_type,
+        ..CachedModule::default()
+      })
+    }
+  }
+
+  impl SpecifierHandler for MockSpecifierHandler {
+    fn fetch(&mut self, specifier: ModuleSpecifier) -> FetchFuture {
+      Box::pin(future::ready(self.get_cache(specifier)))
+    }
+    fn get_build_info(
+      &self,
+      specifier: &ModuleSpecifier,
+      _cache_type: &EmitType,
+    ) -> Result<Option<TextDocument>, AnyError> {
+      Ok(self.build_info.get(specifier).cloned())
+    }
+    fn set_cache(
+      &mut self,
+      specifier: &ModuleSpecifier,
+      cache_type: &EmitType,
+      code: TextDocument,
+      maybe_map: Option<TextDocument>,
+    ) -> Result<(), AnyError> {
+      self.cache_calls.push((
+        specifier.clone(),
+        cache_type.clone(),
+        code,
+        maybe_map,
+      ));
+      Ok(())
+    }
+    fn set_types(
+      &mut self,
+      specifier: &ModuleSpecifier,
+      types: String,
+    ) -> Result<(), AnyError> {
+      self.types_calls.push((specifier.clone(), types));
+      Ok(())
+    }
+    fn set_build_info(
+      &mut self,
+      specifier: &ModuleSpecifier,
+      cache_type: &EmitType,
+      build_info: TextDocument,
+    ) -> Result<(), AnyError> {
+      self
+        .build_info
+        .insert(specifier.clone(), build_info.clone());
+      self.build_info_calls.push((
+        specifier.clone(),
+        cache_type.clone(),
+        build_info,
+      ));
+      Ok(())
+    }
+    fn set_deps(
+      &mut self,
+      specifier: &ModuleSpecifier,
+      dependencies: DependencyMap,
+    ) -> Result<(), AnyError> {
+      self.deps_calls.push((specifier.clone(), dependencies));
+      Ok(())
+    }
+    fn set_version(
+      &mut self,
+      specifier: &ModuleSpecifier,
+      version: String,
+    ) -> Result<(), AnyError> {
+      self.version_calls.push((specifier.clone(), version));
+      Ok(())
+    }
+  }
 
   #[test]
   fn test_get_version() {
